+++ scripts/wdio/clean-screenshots-cli.js
@@ -8,8 +8,8 @@ const packageJson = require('../../package.json');
 // Parse process arguments
 commander
   .version(packageJson.version)
-  .option('--config <path>', 'The wdio config for the tests. Alias for <config>.', undefined)
-  .option('--removeReference <path>', 'runs a certain spec file, can be combined with --suite - overrides specs piped from stdin', undefined)
+  .option('--config <path>', 'The wdio config path for the tests', undefined)
+  .option('--removeReference', 'Whether or not to remove the reference screenshots', false)
   .parse(process.argv);
 
 const {